+++ app/src/main/java/com/bignerdranch/android/activityplanner/model/BusinessCategory.kt
@@ -9,5 +9,5 @@ import androidx.room.Entity
 )
 data class BusinessCategory(
      val businessId: String,
-     @ColumnInfo(name = "categoryId", index = true) val categoryId: Long
+     @ColumnInfo(name = "categoryId", index = true) val categoryId: Int
 )
